почищен код от ненужных элементов
