+++ src/model/destination-model.js
@@ -15,44 +15,4 @@ export default class DestinationsModel extends Observable {
   set destinations (destinations) {
     this.#destinations = destinations;
   }
-
-  updateDestination(updateType, update) {
-    const index = this.#destinations.findIndex((destination) => destination.id === update.id);
-
-    if (index === -1) {
-      throw new Error('Can\'t update unexisting task');
-    }
-
-    this.#destinations = [
-      ...this.#destinations.slice(0, index),
-      update,
-      ...this.#destinations.slice(index + 1),
-    ];
-
-    this._notify(updateType, update);
-  }
-
-  addDestination(updateType, update) {
-    this.#destinations = [
-      update,
-      ...this.#destinations,
-    ];
-
-    this._notify(updateType, update);
-  }
-
-  deleteDestination(updateType, update) {
-    const index = this.#destinations.findIndex((destination) => destination.id === update.id);
-
-    if (index === -1) {
-      throw new Error('Can\'t delete unexisting task');
-    }
-
-    this.#destinations = [
-      ...this.#destinations.slice(0, index),
-      ...this.#destinations.slice(index + 1),
-    ];
-
-    this._notify(updateType);
-  }
 }
