+++ src/model/offers-model.js
@@ -15,44 +15,4 @@ export default class OffersModel extends Observable {
   set offers (offers) {
     this.#offers = offers;
   }
-
-  updateOffer(updateType, update) {
-    const index = this.#offers.findIndex((offer) => offer.id === update.id);
-
-    if (index === -1) {
-      throw new Error('Can\'t update unexisting task');
-    }
-
-    this.#offers = [
-      ...this.#offers.slice(0, index),
-      update,
-      ...this.#offers.slice(index + 1),
-    ];
-
-    this._notify(updateType, update);
-  }
-
-  addOffer(updateType, update) {
-    this.#offers = [
-      update,
-      ...this.#offers,
-    ];
-
-    this._notify(updateType, update);
-  }
-
-  deleteOffer(updateType, update) {
-    const index = this.#offers.findIndex((offer) => offer.id === update.id);
-
-    if (index === -1) {
-      throw new Error('Can\'t delete unexisting task');
-    }
-
-    this.#offers = [
-      ...this.#offers.slice(0, index),
-      ...this.#offers.slice(index + 1),
-    ];
-
-    this._notify(updateType);
-  }
 }
